modificando el nombre de las rutas
