--- app/controllers/api/open_library_controller.rb
@@ -1,9 +1,0 @@
-module Api
-  class OpenLibraryController < ApplicationController
-    def show
-      @book = OpenLibraryService.new(params[:isbn]).book_info
-      render json: @book[:result], status: @book[:status?]
-    end
-
-  end
-end
