+++ app/models/book.rb
@@ -15,4 +15,5 @@ class Book < ApplicationRecord
   end
 
   has_many :rent
+
 end
