+++ app/serializers/user_serializer.rb
@@ -1,7 +1,8 @@
 class UserSerializer < Panko::Serializer
-    attributes :id, :first_name, :last_name, :email, :rents_counter
+  attributes :id, :first_name, :last_name, :email, :rents_counter
 
-    def rents_counter
-        "#{object.rent.count}"
-    end
+  def rents_counter
+    object.rent.count.to_s
+  end
+
 end
