+++ spec/models/book_suggestion_spec.rb
@@ -80,5 +80,4 @@ RSpec.describe BookSuggestion, type: :model do
       end
     end
   end
-  
 end
